requirements.txt
python-dotenv
python-gitlab
tqdm
